removing unit tests for MySQL connection
We don't have access to the service account credentials during the build process. Accessing these credentials is necessary to test the connection to the MySQL server therefore for the time being we will not test the connection. Instead, I am shifting towards being able to mock the DB connection so we can still run unit tests and the such.