added neo4j unit test runner
Neo4J provides a test harness that is able to start a local server which is launched during unit tests. Note, there is currently the problem that these tests ~45 seconds to run which is too long for me liking. A possible solution is integrating the nosql-unit neo4j in-memory instance, but my experimentation with that took to long to setup especially in conjuction with spring boot