FIX: Adding optional parameter to be able to set panel to a new artifact panel
I noticed that creating an artifact from the context menu while one was selected led to the artifact panel thinking it was editing the artifact. Therefore, I added an optional parameter to set the panel to be a new artifact creator panel by clearing the currently selected artifact is parameter is passed.