REFACTOR: lazy loading trainer to enable faster response time for prediction endpoint.
When running the prediction job on the server the trainer used to be initialized before returning the `outputPath`. However, this required waiting for the download of the model weights which could take a long time. Therefore, trainer is loaded once thread starts allowing the actual HTTP response to be sent back whilst loading the weights.